Extract PhpDocsResolver from NodeScopeResolver
PhpClassReflectionExtension no longer needs the whole NodeScopeResolver -
it injects the PhpDocsResolver it actually uses.
